# GeoJSON to/from custom struct using serde