--- a/benches/parse.rs
+++ b/benches/parse.rs
@@ -3,9 +3,9 @@ use geojson::GeoJson;
 use std::io::BufReader;
 
 fn parse_feature_collection_benchmark(c: &mut Criterion) {
-    c.bench_function("parse (countries.geojson)", |b| {
-        let geojson_str = include_str!("../tests/fixtures/countries.geojson");
+    let geojson_str = include_str!("../tests/fixtures/countries.geojson");
 
+    c.bench_function("parse (countries.geojson)", |b| {
         b.iter(|| {
             let _ = black_box({
                 match geojson_str.parse::<geojson::GeoJson>() {
@@ -19,17 +19,52 @@ fn parse_feature_collection_benchmark(c: &mut Criterion) {
     });
 
     c.bench_function("FeatureIter (countries.geojson)", |b| {
-        let geojson_str = include_str!("../tests/fixtures/countries.geojson");
-
         b.iter(|| {
             let feature_iter =
                 geojson::FeatureIterator::new(BufReader::new(geojson_str.as_bytes()));
             let _ = black_box({
                 let mut count = 0;
-                for _ in feature_iter {
+                for feature in feature_iter {
+                    let _ = feature.unwrap();
                     count += 1;
                 }
-                assert_eq!(count, 184);
+                assert_eq!(count, 180);
+            });
+        });
+    });
+
+    c.bench_function("FeatureReader::features (countries.geojson)", |b| {
+        b.iter(|| {
+            let feature_reader =
+                geojson::FeatureReader::from_reader(BufReader::new(geojson_str.as_bytes()));
+            let _ = black_box({
+                let mut count = 0;
+                for feature in feature_reader.features() {
+                    let _ = feature.unwrap();
+                    count += 1;
+                }
+                assert_eq!(count, 180);
+            });
+        });
+    });
+
+    c.bench_function("FeatureReader::deserialize (countries.geojson)", |b| {
+        b.iter(|| {
+            #[derive(serde::Deserialize)]
+            struct Country {
+                geometry: geojson::Geometry,
+                name: String,
+            }
+            let feature_reader =
+                geojson::FeatureReader::from_reader(BufReader::new(geojson_str.as_bytes()));
+
+            let _ = black_box({
+                let mut count = 0;
+                for feature in feature_reader.deserialize::<Country>().unwrap() {
+                    let _ = feature.unwrap();
+                    count += 1;
+                }
+                assert_eq!(count, 180);
             });
         });
     });