--- a/benches/parse.rs
+++ b/benches/parse.rs
@@ -1,5 +1,8 @@
-use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use geojson::de::deserialize_geometry;
 use geojson::GeoJson;
+
+use criterion::{black_box, criterion_group, criterion_main, Criterion};
+
 use std::io::BufReader;
 
 fn parse_feature_collection_benchmark(c: &mut Criterion) {
@@ -54,6 +57,33 @@ fn parse_feature_collection_benchmark(c: &mut Criterion) {
             });
         });
     });
+
+    #[cfg(feature="geo-types")]
+    c.bench_function(
+        "FeatureReader::deserialize to geo-types (countries.geojson)",
+        |b| {
+            b.iter(|| {
+                #[allow(unused)]
+                #[derive(serde::Deserialize)]
+                struct Country {
+                    #[serde(deserialize_with = "deserialize_geometry")]
+                    geometry: geo_types::Geometry,
+                    name: String,
+                }
+                let feature_reader =
+                    geojson::FeatureReader::from_reader(BufReader::new(geojson_str.as_bytes()));
+
+                let _ = black_box({
+                    let mut count = 0;
+                    for feature in feature_reader.deserialize::<Country>().unwrap() {
+                        let _ = feature.unwrap();
+                        count += 1;
+                    }
+                    assert_eq!(count, 180);
+                });
+            });
+        },
+    );
 }
 
 fn parse_geometry_collection_benchmark(c: &mut Criterion) {