--- a/benches/serialize.rs
+++ b/benches/serialize.rs
@@ -1,4 +1,5 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use geojson::{de::deserialize_geometry, ser::serialize_geometry};
 
 fn serialize_feature_collection_benchmark(c: &mut Criterion) {
     let geojson_str = include_str!("../tests/fixtures/countries.geojson");
@@ -39,6 +40,38 @@ fn serialize_feature_collection_benchmark(c: &mut Criterion) {
             });
         });
     });
+
+    #[cfg(feature="geo-types")]
+    c.bench_function(
+        "serialize custom struct to geo-types (countries.geojson)",
+        |b| {
+            #[derive(serde::Serialize, serde::Deserialize)]
+            struct Country {
+                #[serde(
+                    serialize_with = "serialize_geometry",
+                    deserialize_with = "deserialize_geometry"
+                )]
+                geometry: geo_types::Geometry,
+                name: String,
+            }
+            let features =
+                geojson::de::deserialize_feature_collection_str_to_vec::<Country>(geojson_str)
+                    .unwrap();
+            assert_eq!(features.len(), 180);
+
+            b.iter(|| {
+                black_box({
+                    let geojson_string =
+                        geojson::ser::to_feature_collection_string(&features).unwrap();
+                    // Sanity check that we serialized a long string of some kind.
+                    //
+                    // Note this is slightly shorter than the GeoJson round-trip above because we drop
+                    // some fields, like foreign members
+                    assert_eq!(geojson_string.len(), 254908);
+                });
+            });
+        },
+    );
 }
 
 criterion_group!(benches, serialize_feature_collection_benchmark);