--- a/benches/to_geo_types.rs
+++ b/benches/to_geo_types.rs
@@ -4,6 +4,7 @@ fn benchmark_group(c: &mut Criterion) {
     let geojson_str = include_str!("../tests/fixtures/countries.geojson");
     let geojson = geojson_str.parse::<geojson::GeoJson>().unwrap();
 
+    #[cfg(feature="geo-types")]
     c.bench_function("quick_collection", move |b| {
         b.iter(|| {
             let _: Result<geo_types::GeometryCollection<f64>, _> =