--- a/src/conversion/mod.rs
+++ b/src/conversion/mod.rs
@@ -63,12 +63,10 @@ macro_rules! assert_almost_eq {
     }};
 }
 
-
 macro_rules! try_from_owned_value {
     ($to:ty) => {
         #[cfg_attr(docsrs, doc(cfg(feature = "geo-types")))]
-        impl<T: CoordFloat> TryFrom<geometry::Value> for $to
-        {
+        impl<T: CoordFloat> TryFrom<geometry::Value> for $to {
             type Error = Error;
 
             fn try_from(value: geometry::Value) -> Result<Self> {
@@ -77,7 +75,6 @@ macro_rules! try_from_owned_value {
         }
     };
 }
-
 
 pub(crate) mod from_geo_types;
 pub(crate) mod to_geo_types;