… CHANGES.md
… Cargo.toml
… examples/deserialize.rs
… examples/deserialize_to_geo_types.rs