@metasim 2022-08-18
This feature would be brilliant!