@rmanoka 2022-08-22
This is indeed a fantastic idea: really flexible and ergonomic conversion to/from geojson!  Thanks @michaelkirk