@rmanoka 2022-08-25
The ser. support is awesome!  It should address #176 too, right?

closes #176