@michaelkirk 2022-08-25
> The ser. support is awesome! It should address https://github.com/georust/geojson/issues/176 too, right?

Unfortunately, it's probably not much of a performance improvement at this point. In fact, it might be worse in some ways.

For one, there's this really unfortunate `PERF` note here: https://github.com/georust/geojson/pull/199/files#diff-02292e8d776b8c5c924b5e32f227e772514cb68b37f3f7b384f02cdb6717a181R319 

...where, in order to "move" a structs fields into the output's "properties" field, we roundtrip each of your structs to a `serde_json::Object`.

We do this round trip with only one feature at a time, not the entire feature collection at once, so its memory impact should be limited, but still, it's pretty gross. With the time I had, I wasn't able to come up with a better solution.

But the bigger problem might be that (currently anyway) serialization requires a `&[T: Deserialize]`, so you already need your entire slice of structs to be in memory.