@rmanoka 2022-08-25
The conversion to `JsonObject` at `Feature` level is still okay.  Currently, our serialize impl.:

```
impl Serialize for FeatureCollection {
    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
    where
        S: Serializer,
    {
        JsonObject::from(self).serialize(serializer)
    }
}
```

converts the whole feature collection into a much larger representation before writing.  Typically the user is already storing the representation in memory, and the conversion takes a lot (~10x) more memory.  This PR nicely side-steps this issue, right?