@michaelkirk 2022-08-25
> Could we get a short doc string here so it's visible in the docs?


I've done this in https://github.com/georust/geojson/pull/199/commits/493463eb9bea7856ae4f04656cdfa06c228956bb

And then I realized that the short doc string was basically indistinguishable from `FeatureIterator`. To avoid the confusion to users of having similar sounding things,  I decided to deprecate the public `FeatureIterator` in  https://github.com/georust/geojson/pull/199/commits/269e29d2bdde946ebf607016ec58d1260dc7e42e, in favor of accessing it opaquely (as an impl Iterator) on `FeatureReader::features`. 

The thought is that we wouldn't delete it, but eventually make it private once the deprecation cycle has run its course.

So to recap the new suggested usage:

`FeatureReader::features()` gives you an iterator over `Feature`s, just like `FeatureIterator::new(io_stream)` does today.

And then:

`FeatureReader::.deserialize()` gives you an iterator over your custom serde struct.

This is somewhat inspired by the approach taken by [`csv::Reader::records()`](https://docs.rs/csv/latest/csv/struct.Reader.html#method.records) vs. [`csv::Reader::deserialize()`](https://docs.rs/csv/latest/csv/struct.Reader.html#method.deserialize)


WDYT?