@michaelkirk 2022-08-25
> should we call out this functionality early on in the "main" docs?

Done in https://github.com/georust/geojson/pull/199/commits/413b959add564933ef7e1ee76da1a1ed3836834b